requirements.txt
numpy
opencv-python
pandas
Pillow
torch
torchvision
